requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
mysqlclient>=2.2,<3.0
# Necessário em Windows para fusos horários
TZData>=2024.1
